requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
greenlet==3.1.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
missingno==0.5.2
numpy==1.24.4
packaging==24.2
pandas==2.0.3
patsy==0.5.6
pillow==11.0.0
plotly==5.24.1
psycopg2==2.9.9
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.29
statsmodels==0.14.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
